# nashikbackend-main/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EvidenceShield AI API
  description: Tamper-Evident Zero-Trust Judicial & Forensic Document Management System
  version: 1.0.0-s2
servers:
  - url: /api/v1
    description: Current Base API Server
paths:
  /auth/session:
    get:
      summary: Get Current Authenticated Session
      operationId: getCurrentSession
      responses:
        '200':
          description: Active user session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/mfa/verify:
    post:
      summary: Verify TOTP MFA for Fresh Session
      operationId: verifyMfa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: MFA verified, freshness window updated to 10 minutes
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: boolean }
                  mfa_verified_at: { type: string, format: date-time }
        '400':
          $ref: '#/components/responses/BadRequestError'
...
  /cases/{id}/alerts:
    get:
      summary: List Case Integrity Alerts
      operationId: listAlerts
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IntegrityAlert'
  /alerts/{id}/review:
    post:
      summary: Review and Set Disposition for Alert
      operationId: reviewAlert
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertReviewRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrityAlert'
...
  /cases/{id}/exports:
    post:
      summary: Generate Certified Judicial Export Package
      operationId: generateExport
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ZIP bundle metadata and download token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPackageResponse'
components:
  schemas:
    SessionInfo:
      type: object
      required: [user_id, username, role, agency, mfa_verified, mfa_fresh]
      properties:
        user_id: { type: string }
        username: { type: string }
        role:
          type: string
          enum: [POLICE, FORENSIC_LAB, PROSECUTOR, SYSTEM_ADMIN, AUDITOR]
        agency: { type: string }
        mfa_verified: { type: boolean }
        mfa_fresh: { type: boolean }
...
    IntegrityAlert:
      type: object
      required: [alert_id, version_id, case_id, alert_type, failure_stage, status, created_at]
      properties:
        alert_id: { type: string, format: uuid }
        version_id: { type: string, format: uuid }
        case_id: { type: string, format: uuid }
        alert_type: { type: string }
        failure_stage: { type: string }
        error_details: { type: string }
        status:
          type: string
          enum: [OPEN, INVESTIGATING, CONFIRMED_TAMPERED, DISMISSED_FALSE_POSITIVE]
        disposition_notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    AlertReviewRequest:
      type: object
      required: [status, disposition_notes]
      properties:
        status:
          type: string
          enum: [CONFIRMED_TAMPERED, DISMISSED_FALSE_POSITIVE]
        disposition_notes: { type: string }
...
    ExportPackageResponse:
      type: object
      required: [export_id, case_id, download_url, expires_at]
      properties:
        export_id: { type: string, format: uuid }
        case_id: { type: string, format: uuid }
        download_url: { type: string }
        expires_at: { type: string, format: date-time }
    ErrorEnvelope:
      type: object
      required: [type, title, status, detail, instance, code, timestamp]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        timestamp: { type: string, format: date-time }
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnauthorizedError:
      description: Authentication required or invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...